ci(profiling): refactor pytorch profiling tests into their own file.
